+++ src/Ema/Class.hs
@@ -1,29 +1,31 @@
 {-# LANGUAGE DefaultSignatures #-}
-{-# LANGUAGE FunctionalDependencies #-}
 
 module Ema.Class where
 
 -- | Enrich a model to work with Ema
-class Ema model route | route -> model where
+class Ema r where
+  type ModelFor r :: Type
+
   -- | Get the filepath on disk corresponding to this route.
-  encodeRoute :: model -> route -> FilePath
+  encodeRoute :: ModelFor r -> r -> FilePath
 
   -- | Decode a filepath on disk into a route.
-  decodeRoute :: model -> FilePath -> Maybe route
+  decodeRoute :: ModelFor r -> FilePath -> Maybe r
 
   -- | All routes in the site
   --
   -- The `gen` command will generate only these routes. On live server, this
   -- function is never used.
-  allRoutes :: model -> [route]
-  default allRoutes :: (Bounded route, Enum route) => model -> [route]
+  allRoutes :: ModelFor r -> [r]
+  default allRoutes :: (Bounded r, Enum r) => ModelFor r -> [r]
   allRoutes _ = [minBound .. maxBound]
 
 -- | The unit model is useful when using Ema in pure fashion (see
 -- @Ema.runEmaPure@) with a single route (index.html) only.
-instance Ema () () where
-  encodeRoute () () = []
+instance Ema () where
+  type ModelFor () = ()
+  encodeRoute () () = "index.html"
   decodeRoute () = \case
-    [] -> Just ()
+    "index.html" -> Just ()
     _ -> Nothing
   allRoutes () = one ()