+++ src/Ema/Example/Ex02_Basic.hs
@@ -5,7 +5,6 @@ import Control.Concurrent (threadDelay)
 import Data.LVar qualified as LVar
 import Ema (Ema (..))
 import Ema qualified
-import Ema.CLI qualified as CLI
 import Ema.Example.Common (tailwindLayout)
 import Text.Blaze.Html5 ((!))
 import Text.Blaze.Html5 qualified as H
@@ -18,7 +17,8 @@ data Route
 
 newtype Model = Model {unModel :: Text}
 
-instance Ema Model Route where
+instance Ema Route where
+  type ModelFor Route = Model
   encodeRoute _model =
     \case
       Index -> "index.html"
@@ -31,7 +31,7 @@ instance Ema Model Route where
 main :: IO ()
 main = do
   void $
-    Ema.runEma (\_act m -> Ema.AssetGenerated Ema.Html . render m) $ \act model -> do
+    Ema.runEma (\_act m -> Ema.AssetGenerated Ema.Html . render m) $ \_act model -> do
       LVar.set model $ Model "Hello World. "
       liftIO $ threadDelay maxBound