+++ src/Ema/Generate.hs
@@ -6,7 +6,7 @@ module Ema.Generate where
 import Control.Exception (throw)
 import Control.Monad.Logger
 import Ema.Asset (Asset (..))
-import Ema.Class (Ema (allRoutes, encodeRoute))
+import Ema.Class (Ema (..))
 import System.Directory (copyFile, createDirectoryIfMissing, doesDirectoryExist, doesFileExist, doesPathExist)
 import System.FilePath (takeDirectory, (</>))
 import System.FilePattern.Directory (getDirectoryFiles)
@@ -16,22 +16,25 @@ log :: MonadLogger m => LogLevel -> Text -> m ()
 log = logWithoutLoc "Generate"
 
 generate ::
-  forall model route m.
+  forall r m.
   ( MonadIO m,
     MonadUnliftIO m,
     MonadLoggerIO m,
-    Ema model route,
+    Ema r,
+    Show r,
     HasCallStack
   ) =>
   FilePath ->
-  model ->
-  (model -> route -> Asset LByteString) ->
+  ModelFor r ->
+  (ModelFor r -> r -> Asset LByteString) ->
   -- | List of generated files.
   m [FilePath]
 generate dest model render = do
   unlessM (liftIO $ doesDirectoryExist dest) $ do
     error $ "Destination does not exist: " <> toText dest
   let routes = allRoutes model
+  when (null routes) $
+    error "allRoutes is empty; nothing to generate"
   log LevelInfo $ "Writing " <> show (length routes) <> " routes"
   let (staticPaths, generatedPaths) =
         lefts &&& rights $